requirements.txt
numpy==2.1.3
pandas
polars
matplotlib
pyod

langchain-core
langchain-community
langchain-chroma
langchain-cohere
langchain-text-splitters 
langgraph
langchain-groq
langchain-openai
langchain-huggingface
langchain-together

openpyxl

